Edit: js_processor.py and related
